requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfminer.six==20221105
pandas==2.2.1
openpyxl==3.1.2
